Add campus-related procedures and models for retrieval and registration
